requirements/base.txt
pyyaml>=6.0.2,<7.0.0
pyperclip>=1.9.0,<2.0.0
undetected-chromedriver==3.5.3
bs4>=0.0.2,<0.0.3
fake-useragent>=2.1.0,<3.0.0
stem>=1.8.2,<2.0.0
icecream
colorlog
Flask==2.2.5